+++ Gemfile
@@ -26,6 +26,7 @@ gem 'faker', :git => 'https://github.com/faker-ruby/faker.git', :branch => 'mast
 gem 'will_paginate', '3.1.8'
 gem 'bootstrap-will_paginate'
 gem 'turbolinks_render'
+# gem 'carrierwave'
 
 group :development do
   gem 'web-console', '>= 3.3.0'
