+++ app/assets/stylesheets/antiques/_card_gridv2.scss
@@ -15,19 +15,23 @@
     align-items: center;
   }
   #new-antique-btn {
+    @include themed() {
+      color: t('color-2');
+      background-color: t('background-color-elements');
+    }
     transition: 0.3s all ease;
     float: right;
     text-decoration: none;
-    color: rgba(40, 40, 40, 0.7);
     padding: 5px 15px;
     text-align: center;
     line-height: 1.3rem;
     border-radius: 3px;
-    background-color: rgba(40, 40, 40, 0.1);
     &:hover {
       text-decoration: underline;
-      color: black;
-      background-color: rgba(40, 40, 40, 0.2);
+      @include themed() {
+        color: t('color-2-hover');
+        background-color: t('background-color-elements-hover');
+      }
     }
   }
 }
