+++ app/assets/stylesheets/antiques/_fullscreen_carousel_modal.scss
@@ -1,3 +1,8 @@
+.fs-carousel-haut {
+  height: 10vh;
+  display: flex;
+}
+
 #fullscreen-carousel-dialog {
   width: 100%;
   height: 100%;
@@ -45,21 +50,50 @@
 }
 
 #fs-carousel-caption {
-  bottom: 0;
+  bottom: -15vh;
   background-color: rgba(0, 0, 0, 0.3);
   padding: 20px;
   width: 100vw;
   left: 0;
   right: 0;
+  height: 15vh;
 }
+
+.carousel-caption {
+  p {
+    height: calc(100% - 1.25rem);
+    overflow-y: scroll;
+    margin-bottom: 0;
+    width: 80vw;
+    max-width: 950px;
+    margin: 0 auto;
+  }
+}
+
+  ::-webkit-scrollbar {
+    display: none;
+    -ms-overflow-style: none;  /* IE and Edge */
+    scrollbar-width: none;  /* Firefox */
+    width: 0px;
+    background: transparent; /* make scrollbar transparent */
+  }
 
 #carousel-fs-inner {
-  height: 70vh;
+  height: 85vh;
 }
 
-.fs-carousel-haut-bas {
-  height: 15vh;
+.fs-carousel-bas {
+  height: 5vh;
   display: flex;
+  flex-direction: column;
+  justify-content: center;
+}
+
+
+.carousel-indicators {
+  // margin-bottom: 0;
+  margin: 0 auto;
+  bottom: unset;
 }
 
 .fs-carousel-indicators {
