+++ app/assets/stylesheets/application.scss
@@ -13,3 +13,4 @@
 @import "components/index";
 @import "pages/index";
 @import "antiques/index";
+
