+++ app/assets/stylesheets/config/_dark_mode.scss
@@ -17,7 +17,7 @@ $themes: (
     color-2: hsla(0, 0%, 10%, 1),
     color-2-hover: hsla(0, 0%, 25%, 1),
     background-color-main: hsla(0, 0%, 100%, 1),
-    background-color-elements: hsla(0, 0%, 99%, 1),
+    background-color-elements: hsla(0, 0%, 98%, 1),
     background-color-elements-hover: hsla(0, 0%, 89%, 1)
 
   ),
