+++ app/javascript/packs/application.js
@@ -6,11 +6,10 @@ import 'bootstrap';
 
 import { themeToggle }                                                from '../components/theme_toggle';
 import { closeShowModalWhenModify }                                   from '../pages/index';
-import { openCarouselOnClickedElement, navigateFSCarouselWithArrows } from '../components/fullscreencarousel';
 import { passSrcToVideoModal }                                        from '../components/videomodal';
 import { initSwiperHome }                                             from '../components/homecarousel';
 import { initMap }                                                    from '../components/googlemap';
-import { openCarouselOnClickedElementV2 }                             from '../components/fullscreen_carouselv2';
+import { openCarouselOnClickedElement, navigateFSCarouselWithArrows } from '../components/fullscreen_carousel';
 
 $(document).on('turbolinks:load', function() {
 
@@ -20,7 +19,6 @@ $(document).on('turbolinks:load', function() {
   closeShowModalWhenModify();
   passSrcToVideoModal();
   openCarouselOnClickedElement();
-  openCarouselOnClickedElementV2();
   initSwiperHome();
   initMap();
