+++ app/views/antiques/_new.html.erb
@@ -8,6 +8,9 @@
         Ajouter une photo <em> (10.5 Mb max.)</em>
         <span style="display:none;">
           <%= f.input :photo, as: :file, required: true, name: 'Photo' %>
+          <%#= f.input :photo, as: :file, name: 'Photo', input_html: { multiple: true } %>
+          <%#= f.input :photo, as: :file, input_html: { class: 'hidden', id: 'photo-input', multiple: true }, label_html: { class: 'upload-photo'}, label: '📷 Upload a photo' %>
+          <%#= image_tag "", width: 300, class: "hidden", id: "photo-preview" %>
         </span>
       </label>
       <hr>
