+++ app/views/shared/_navbar_v2.html.erb
@@ -27,11 +27,21 @@
               <% end %>
             </div>
           <% end %>
+          <% if user_signed_in? %>
+            <div class="dropdown-menu-elem-flex">
+              <%= link_to "Nouvelle Galerie", "#", class: "dropdown-item-nav photo-cat un", data: { toggle: "modal", target: "#newCategoryModal" }, id: "category-navbar" %>
+            </div>
+          <% end %>
         </div>
       </div>
-      <div class="navbar-element dropdown">
+      <div class="navbar-element">
         <%= link_to "Contact", "#", class: "navbar-element-text", data: { toggle: "modal", target: "#newContactModal" }, id: "contact-navbar" %>
       </div>
+      <% if user_signed_in? %>
+        <div class="navbar-element">
+          <%= link_to "Déconnexion", destroy_user_session_path, method: :delete, class: "navbar-element-text", id: "log-out-link" %>
+        </div>
+      <% end %>
       <!-- Theme switch -->
       <div class="theme-items">
         <%= link_to "", root_path(theme: "#{ current_opposite_theme }"), id: "theme-link" %>
@@ -49,3 +59,37 @@
 </div>
 
 <%= render 'shared/contact_modal' %>
+  <!-- Modal New Category -->
+  <div class="modal fade" id="newCategoryModal" tabindex="-1" role="dialog" aria-labelledby="newCategoryModalTitle" aria-hidden="true">
+    <div class="modal-dialog modal-dialog-centered" role="document">
+      <div class="modal-content">
+        <div class="modal-header">
+          <h5 class="modal-title" id="exampleModalLongTitle">Création d'une nouvelle galerie</h5>
+          <button id="new-cat-modal-close-btn" type="button" class="close" data-dismiss="modal" aria-label="Close">
+            <span aria-hidden="true">&times;</span>
+          </button>
+        </div>
+        <div class="new-category-modal-body">
+          <%= render 'categories/new', category: Category.new %>
+        </div>
+      </div>
+    </div>
+  </div>
+
+    <% Category.all.order("created_at ASC").each do |cat| %>
+    <div class="modal fade" id="editCategoryModal<%= cat.id %>" tabindex="-1" role="dialog" aria-labelledby="editCategoryModalTitle" aria-hidden="true">
+      <div class="modal-dialog modal-dialog-centered" role="document">
+        <div class="modal-content">
+          <div class="modal-header">
+            <h5 class="modal-title" id="exampleModalLongTitle">Modifier <%= cat.title %></h5>
+            <button id="edit-cat-modal-close-btn" type="button" class="close" data-dismiss="modal" aria-label="Close">
+              <span aria-hidden="true">&times;</span>
+            </button>
+          </div>
+          <div class="edit-category-modal-body">
+            <%= render 'categories/edit', category: cat%>
+          </div>
+        </div>
+      </div>
+    </div>
+  <% end %>
